MissionBugfixes
Backup of most missions which conflict with either Vanilla or BT10. Others which conflicted have been mostly checked and seemed to be down to bugs.
Spain, Portugal, Netherlands have all been cut for now, Amyz, can you put them back in and tell me which are working properly and are updated properly?