Bugfixes, Reimplemetnations everytihgn
Adding in all outside IR mod-files and vanilla files back into IR to solve some bugissues.
Plan is to not remove these files at all, but rather, consolidate them, inactivate them and other things that are necessary to make the mod run smooth and balanced. This will require a lot of work in the future. But it will save a lot of work bugfixing things that are broken.